ai-service/requirements.txt
# FastAPI & Server
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# AI & LLM
google-generativeai==0.8.4
langchain==0.3.14
langchain-google-genai==2.0.8

# NLP
spacy==3.8.4

# ML & Data
scikit-learn==1.6.1
pandas==2.2.3
numpy==2.2.2

# Image Processing
Pillow==11.1.0

# Database
sqlalchemy==2.0.37
psycopg2-binary==2.9.10
pymongo==4.10.1
redis==5.2.1

# Utils
pydantic==2.10.5
pydantic-settings==2.7.1
python-dotenv==1.0.1
httpx==0.28.1
